+++ testing/test_cli.py
@@ -225,7 +225,8 @@ def test_cli_read_telemetry_success(caplog):
     assert (
         "Sending message to udp://255.255.255.255:60110\n"
         "$MLHDT,235.0,T*27\r\n"
-        "$MLVWR,154.0,L,11.06,N,5.69,M,20.48,K*64" in caplog.messages
+        "$MLVWR,154.0,L,11.06,N,5.69,M,20.48,K*64\r\n"
+        "$MLXDR,A,-60.0,D,PTCH#CAL,A,30.0,D,ROLL#CAL*75" in caplog.messages
     )