+++ testing/test_examples.py
@@ -15,5 +15,6 @@ def test_telemetry_nmea0183_demo(caplog):
     assert (
         "Sending message to udp://255.255.255.255:60110\n"
         "$MLHDT,235.0,T*27\r\n"
-        "$MLVWR,154.0,L,11.06,N,5.69,M,20.48,K*64" in caplog.messages
+        "$MLVWR,154.0,L,11.06,N,5.69,M,20.48,K*64\r\n"
+        "$MLXDR,A,-60.0,D,PTCH#CAL,A,30.0,D,ROLL#CAL*75" in caplog.messages
     )